[review comments] Refactor `Timeouts` to vocabulary type in Cython
