+++ tests/unit/test_session.py
@@ -69,11 +69,13 @@ def test_session_constructed(ext_cls):
         channel_high_watermark=8000000,
         event_queue_watermarks=(6000000, 7000000),
         stats_dump_interval=30.0,
-        connect_timeout=None,
-        disconnect_timeout=None,
-        open_queue_timeout=60.0,
-        configure_queue_timeout=60.0,
-        close_queue_timeout=60.0,
+        timeouts=Timeouts(
+            connect_timeout=None,
+            disconnect_timeout=None,
+            open_queue_timeout=60.0,
+            configure_queue_timeout=60.0,
+            close_queue_timeout=60.0,
+        ),
         monitor_host_health=False,
         fake_host_health_monitor=None,
     )
@@ -123,11 +125,7 @@ def test_session_constructed_with_timeouts(ext_cls):
         channel_high_watermark=8000000,
         event_queue_watermarks=(6000000, 7000000),
         stats_dump_interval=30.0,
-        connect_timeout=60.0,
-        disconnect_timeout=70.0,
-        open_queue_timeout=80.0,
-        configure_queue_timeout=90.0,
-        close_queue_timeout=100.0,
+        timeouts=timeouts,
         monitor_host_health=False,
         fake_host_health_monitor=None,
     )
@@ -171,11 +169,7 @@ def test_session_constructed_with_default_timeouts(ext_cls):
         channel_high_watermark=8000000,
         event_queue_watermarks=(6000000, 7000000),
         stats_dump_interval=30.0,
-        connect_timeout=None,
-        disconnect_timeout=None,
-        open_queue_timeout=None,
-        configure_queue_timeout=None,
-        close_queue_timeout=None,
+        timeouts=timeouts,
         monitor_host_health=False,
         fake_host_health_monitor=None,
     )
@@ -210,11 +204,7 @@ def test_session_default_with_options(ext_cls):
         channel_high_watermark=None,
         event_queue_watermarks=None,
         stats_dump_interval=None,
-        connect_timeout=None,
-        disconnect_timeout=None,
-        open_queue_timeout=None,
-        configure_queue_timeout=None,
-        close_queue_timeout=None,
+        timeouts=Timeouts(),
         monitor_host_health=False,
         fake_host_health_monitor=None,
     )
@@ -266,11 +256,7 @@ def test_session_with_options(ext_cls):
         channel_high_watermark=8000000,
         event_queue_watermarks=(6000000, 7000000),
         stats_dump_interval=30.0,
-        connect_timeout=60.0,
-        disconnect_timeout=70.0,
-        open_queue_timeout=80.0,
-        configure_queue_timeout=90.0,
-        close_queue_timeout=100.0,
+        timeouts=timeouts,
         monitor_host_health=False,
         fake_host_health_monitor=None,
     )
@@ -309,11 +295,13 @@ def test_session_basic_monitor(ext_cls):
         channel_high_watermark=None,
         event_queue_watermarks=None,
         stats_dump_interval=None,
-        connect_timeout=None,
-        disconnect_timeout=None,
-        open_queue_timeout=60.0,
-        configure_queue_timeout=60.0,
-        close_queue_timeout=60.0,
+        timeouts=Timeouts(
+            connect_timeout=None,
+            disconnect_timeout=None,
+            open_queue_timeout=60.0,
+            configure_queue_timeout=60.0,
+            close_queue_timeout=60.0,
+        ),
         monitor_host_health=True,
         fake_host_health_monitor=monitor._monitor,
     )
@@ -344,11 +332,7 @@ def test_session_default_constructed(ext_cls):
         channel_high_watermark=None,
         event_queue_watermarks=None,
         stats_dump_interval=None,
-        connect_timeout=None,
-        disconnect_timeout=None,
-        open_queue_timeout=None,
-        configure_queue_timeout=None,
-        close_queue_timeout=None,
+        timeouts=Timeouts(),
         monitor_host_health=False,
         fake_host_health_monitor=None,
     )
